requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
factory-boy==3.3.0
pytest==7.4.2
pytest-cov==4.1.0
WebTest==3.0.0

# Lint and code style
black==23.10.0
flake8-blind-except==0.2.1
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-isort==6.1.0
flake8==6.1.0
isort==5.12.0
pep8-naming==0.13.3
pytest-playwright

# Local db manipulation
fastparquet
